Reorganize files in src/lib/base
