+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "sketchbook2",
-    "version": "0.0.13",
+    "version": "0.0.14",
     "private": true,
     "scripts": {
         "dev": "vite dev",
