+++ src/config/config.ts
@@ -1,4 +1,4 @@
-import { SortOrder } from '$lib/base/FileLoading/ProjectPresentation';
+import { SortOrder } from '$lib/base/ProjectLoading/ProjectPresentation';
 import { PanelState } from '$lib/base/Util/PanelState';
 
 /**
