+++ src/lib/base/ConfigModels/ParamTypes.ts
@@ -1,9 +1,15 @@
-import { isBooleanParamConfig, type BooleanParamConfig } from './BooleanParamConfig';
-import { isNumberParamConfig, type NumberParamConfig } from './NumberParamConfig';
-import { isFunctionParamConfig, type FunctionParamConfig } from './FunctionParamConfig';
-import { isStringParamConfig, type StringParamConfig } from './StringParamConfig';
-import { isNumericArrayParamConfig, type NumericArrayParamConfig } from './NumericArrayParamConfig';
-import { isFileParamConfig, type FileParamConfig } from './FileParamConfig';
+import { isBooleanParamConfig, type BooleanParamConfig } from './ParamConfigs/BooleanParamConfig';
+import { isNumberParamConfig, type NumberParamConfig } from './ParamConfigs/NumberParamConfig';
+import {
+    isFunctionParamConfig,
+    type FunctionParamConfig
+} from './ParamConfigs/FunctionParamConfig';
+import { isStringParamConfig, type StringParamConfig } from './ParamConfigs/StringParamConfig';
+import {
+    isNumericArrayParamConfig,
+    type NumericArrayParamConfig
+} from './ParamConfigs/NumericArrayParamConfig';
+import { isFileParamConfig, type FileParamConfig } from './ParamConfigs/FileParamConfig';
 
 /*
     File loading result types vary depending on the parameter configuration's read mode
