+++ src/lib/components/ProjectDetailPanel.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import type { ProjectTuple } from '$lib/base/FileLoading/ProjectLoader';
+    import type { ProjectTuple } from '$lib/base/ProjectLoading/ProjectLoader';
     import { settingsStore } from '$lib/base/Util/AppState';
     import Panel from './Panel.svelte';
     import PanelHeader from './PanelHeader.svelte';
