+++ src/lib/components/ProjectListPanel.svelte
@@ -4,7 +4,7 @@
     import { content } from '../../config/content';
     import { settingsStore } from '$lib/base/Util/AppState';
 
-    import type { ProjectConfig } from '$lib/base/ProjectConfig/ProjectConfig';
+    import type { ProjectConfig } from '$lib/base/ConfigModels/ProjectConfig';
     import GroupSelector from './GroupSelector.svelte';
     import Panel from './Panel.svelte';
     import PanelFooter from './PanelFooter.svelte';
