+++ src/routes/+layout.server.ts
@@ -1,6 +1,6 @@
 import type { PageServerLoad } from './$types';
 
-import ProjectLoader from '$lib/base/FileLoading/ProjectLoader';
+import ProjectLoader from '$lib/base/ProjectLoading/ProjectLoader';
 
 export const load: PageServerLoad = (async () => {
     const projects = await ProjectLoader.loadAvailableProjects();
