+++ src/routes/+page.server.ts
@@ -1,5 +1,5 @@
 import { config } from '../config/config';
-import ProjectPresentation, { SortOrder } from '../lib/base/FileLoading/ProjectPresentation';
+import ProjectPresentation, { SortOrder } from '../lib/base/ProjectLoading/ProjectPresentation';
 import type { PageServerLoad } from './$types';
 import { redirect } from '@sveltejs/kit';
