+++ tests/component/GroupSelector.test.ts
@@ -1,8 +1,8 @@
 import { render, fireEvent, screen, cleanup } from '@testing-library/svelte';
 import { describe, it, expect, afterEach } from 'vitest';
-import { type ProjectConfig, ProjectConfigDefaults } from '$lib/base/ProjectConfig/ProjectConfig';
+import { type ProjectConfig, ProjectConfigDefaults } from '$lib/base/ConfigModels/ProjectConfig';
 import GroupSelector from '$lib/components/GroupSelector.svelte';
-import { SortOrder } from '$lib/base/FileLoading/ProjectPresentation';
+import { SortOrder } from '$lib/base/ProjectLoading/ProjectPresentation';
 
 function projectWithGroups(groups: string[], date: number): ProjectConfig {
     const props = {} as ProjectConfig;
