+++ tests/unit/ProjectLoader.test.ts
@@ -6,11 +6,11 @@ import * as Environment from '$app/environment';
 import Project from '$lib/base/Project/Project';
 import ConfigAndSupport from './TestFiles/ConfigAndSupport/ConfigAndSupport';
 import NoConfig from './TestFiles/NoConfig/NoConfig';
-import ProjectLoader from '$lib/base/FileLoading/ProjectLoader';
-import { ProjectConfigDefaults } from '$lib/base/ProjectConfig/ProjectConfig';
-import * as fileProviders from '$lib/base/FileLoading/FileProviders';
-import { ParamType } from '$lib/base/ParamConfig/ParamConfig';
-import ParamValueProvider from '$lib/base/Util/ParamValueProvider';
+import ProjectLoader from '$lib/base/ProjectLoading/ProjectLoader';
+import { ProjectConfigDefaults } from '$lib/base/ConfigModels/ProjectConfig';
+import * as fileProviders from '$lib/base/ProjectLoading/ImportProviders';
+import { ParamType } from '$lib/base/ConfigModels/ParamConfig';
+import ParamValueProvider from '$lib/base/ProjectLoading/ParamValueProvider';
 import FragShaderProject from '$lib/base/Project/FragShaderProject';
 vi.spyOn(ParamValueProvider, 'getValue');
 vi.spyOn(ParamValueProvider, 'setValue');
